[PM-20167] Address audit Issue P
An accross-the-board package update resolves the vulnerable
`tracing-subscriber` instance.

Added `cargo audit` to CI workflows to catch these with regularity going
forward.

For all crates here, the unmaintained `paste` has been replaced by the
drop-in replacement `pastey`, however it has not been removed from the
dependency tree, as a few external dependencies still pull it in.

I consider this acceptable, because it is a relatively simple package,
providing macro tooling, and is unlikely to be subject to
vulnerabilities and need upgrades. Note that cargo audit and RUSTSEC
consider this an 'informative' advisory.