[PM-20162] Address audit Issue K
Extension path construction did not always preserve the invariant that
extensions have <= 255 entries. To address this, extension
constructions, especially with paths of unknown length, should now go
through a new `extension` helper function, that correctly splits the
extension as required. All extension construction that are not trivially
short enough have been replaced with a call to this function.