[PM-20169] Address audit Suggestion 2.
This commit only addresses the specific example in the suggestion.

The suggestion is of course very open-ended, and we are treating it as
an ongoing effort, independent of the push to address audit findings.
Specifically, we have discussed this internally, and agreed to the
following outline plan that is currently being implemented:

- We will be adding code coverage reporting either to on PR checks if
  the time taken is acceptable, or to periodic tests on main.
- We are considering what the best way to surface this coverage
  information is to make sure it is not ignored. Our current sketch is
  to have this be a PR feedback report showing how a PR changes
  coverage metrics.
- This is not yet implemented, but is an independent QA-lead effort.