[PM-20171] Address audit Suggestion 4.
The domain separators in coin structure have been moved to prefix-based
domain separators, and now use a version tag similar to the one we use
in serialization. We have also updated a hash computation for
transaction hashing and hashing for contract addresses to use our tagged
serialization instead of plain serialization, in effect adding a domain
separator to these.

We have not adopted the suggestion of including length prefixes or
structural description prefixes, as this would be a deviation from our
current serialization practice, where structure is a) not
self-describing, but b) prefixed with an identifier that uniquely
defines the format. In keeping with this policy, we will ensure that
domain separators uniquely define the input format expected.

Note that this is a relatively far-reaching breaking change, and will
have knock-on effects for many components, notably Wallet and Compact.