[PM-20163] Further amendments to fix for audit Issues L & B.
On closer review, there are two separate encodings in question here, and
one of these legitimately uses a factor of *16* (not 8 or 32) in both
the specification and implementation. Specifically, state value literals
encoded as field elements are packing more information into a single
numeric value, and this is encoded differently to the *type tag* output
by the `type` opcode and used by the `new` opcode.

The off-by-one error identified in Issues B & L specifically related to
the type tag of Merkle tree values. The field value encoding can
legitimately support a larger space, in part because this is only
encoded into, not decoded from.

As a result, the prior fix for audit Issue L included an adjustment to
the field representation of state value literals, which was not
necessary. This change has been reverted.