[PM-20172] added domain seperators to for dust:proof and zswap:ciphertext
There is no need to add a domain seperator for MerkleTreeNode because of hash
compression, and it's important that there isn't another element being hashed.
Merkle proofs only need collision resistance to work, preimage resistance is
unimportant.

In the communications commitment, the first thing being hashed is randomly selected,
and therefore we don't need a domain separator, as it's already guaranteed to be a
unique input.